Code/LTS/requirements.txt
paddlepaddle==2.1.2
paddlepaddle-gpu==2.1.2 
paddleocr==2.0.6
easyocr==1.4.1
tesseract==4.1.1
opencv-python==4.2.0.32
tensorflow==1.14.0
mmcv-full==1.3.9
mmdet==2.16.0
mmocr==0.3.0
torch==1.9.0
tesserocr==2.5.2
#sudo apt install libleptonica-dev
#sudo apt install libtesseract-dev=4.1.1-2build2

